requirements.txt
# A-Scope Research 金融智能体团队依赖包
# 核心MCP和大模型依赖
langchain-openai==0.1.8
langchain-mcp-adapters==0.1.0
langgraph>=0.1.0

# 配置和环境管理
python-dotenv>=1.0.0
PyYAML>=6.0

# Web界面
streamlit>=1.28.0

# 数据处理和分析
pandas>=2.0.0
numpy>=1.24.0

# HTTP请求和异步处理
httpx>=0.24.0
aiohttp>=3.8.0
requests>=2.31.0

# JSON和数据序列化
orjson>=3.9.0

# 日志和调试
loguru>=0.7.0

# 类型提示
typing-extensions>=4.7.0

# 可选：其他有用的包
# plotly>=5.15.0  # 图表可视化
# matplotlib>=3.7.0  # 图表绘制
# seaborn>=0.12.0  # 统计图表
# scikit-learn>=1.3.0  # 机器学习
# ta-lib>=0.4.0  # 技术分析指标

# 开发和测试依赖（可选）
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.7.0
# flake8>=6.0.0
# mypy>=1.5.0